+++ Plugins/Area/Area.cpp
@@ -71,6 +71,8 @@ Area::Area(Services::ProxyServiceList* services)
     REGISTER(ExportGIT);
     REGISTER(GetTileInfo);
     REGISTER(ExportARE);
+    REGISTER(GetAmbientSoundDay);
+    REGISTER(GetAmbientSoundNight);
 
 #undef REGISTER
 }
@@ -1055,6 +1057,36 @@ ArgumentStack Area::ExportARE(ArgumentStack&& args)
             }
 
             retVal = resGff.WriteGFFFile(alias + ":" + fileName, Constants::ResRefType::ARE);
+        }
+    }
+
+    return Services::Events::Arguments(retVal);
+}
+
+ArgumentStack Area::GetAmbientSoundDay(ArgumentStack&& args)
+{
+    int32_t retVal = 0;
+
+    if (auto *pArea = area(args))
+    {
+        if (pArea->m_pAmbientSound != nullptr)
+        {
+            retVal = pArea->m_pAmbientSound->m_nSoundDayTrack;
+        }
+    }
+
+    return Services::Events::Arguments(retVal);
+}
+
+ArgumentStack Area::GetAmbientSoundNight(ArgumentStack&& args)
+{
+    int32_t retVal = 0;
+
+    if (auto *pArea = area(args))
+    {
+        if (pArea->m_pAmbientSound != nullptr)
+        {
+            retVal = pArea->m_pAmbientSound->m_nSoundNightTrack;
         }
     }