+++ Plugins/Area/Area.hpp
@@ -52,6 +52,8 @@ private:
     ArgumentStack ExportGIT                     (ArgumentStack&& args);
     ArgumentStack GetTileInfo                   (ArgumentStack&& args);
     ArgumentStack ExportARE                     (ArgumentStack&& args);
+    ArgumentStack GetAmbientSoundDay            (ArgumentStack&& args);
+    ArgumentStack GetAmbientSoundNight          (ArgumentStack&& args);
 
     CNWSArea *area(ArgumentStack& args);